rename standard_library -> roc
