+++ compiler/builtins/src/lib.rs
@@ -2,5 +2,5 @@
 // See github.com/rtfeldman/roc/issues/800 for discussion of the large_enum_variant check.
 #![allow(clippy::large_enum_variant)]
 pub mod bitcode;
-pub mod standard_library;
+pub mod roc;
 pub mod std;
